- refactor(app): revert (site) route group; colocate contribute + draft test clients
- Restore about, terms, dashboard, and realm-ranks under src/app/* (no route group).
- Move SupporterTiersShowcase into contribute/_components; fix tier plans import.
- Move DraftTestPageClient into test/draft/_components; fix dynamic import and Convex path.
